kcms/style: Make "Configure Icons and Toolbars" popup fit wide content
Previously it was just hardcoded, which is not appropriate when
variable-length text is involved. To fix it, take into account the
implicit size of the content to determine the popup's size. Also include
a bit of padding so the content doesn't feel completely scrunched up
against the edges of the popup.

BUG: 477303