appiumtests: Introduce a generic plasmoidtest script
A bunch of the appiumtests are simple smoke tests that just check if a
plasmoid loads by trying to find a certain element by name. We can do a
generic version of this by passing the plasmoid ID and the name to find
on the command line, allowing us to drop a bunch of duplicated code.

This adds the generic script and a CMake function to make use of it.