applets/battery: set active status when not in balanced power profile only if profile was not set automatically
Energy savings might be configured to automatically switch to a power profile when on AC/battery/low battery, in which case this should be considered the default configuration, and it might be annoying to make the applet always visible in this case.

Set the active status only if the power profile was selected manually through the applet or is held by an application.