kcms/region_language: repair pt_PT support
this broke because a hack was introduced to change the enumerated name
from pt to pt_PT but that then didn't have a mapping in the glibc table.

to mitigate this and future problems of the sort we now fill the entire
locale name into the mapping table in addition to the presumed language
name. e.g. the table now contains pt=>pt_PT.UTF-8 as well as
pt_PT=>pt_PT.UTF-8

BUG: 478120