applets/mediacontroller: Prevent seek bar from disappearing in Edit Mode
The `enabled` property is transitive, and as such should not be relied
on for the purposes of directly binding opacity.

Test plan: put a media controller applet on a desktop or launch it in
plasmoid viewer; start playing some media; press and hold the applet.
Expected result: Seek bar should not disappear in Edit Mode.