plasma-shutdown: Continue on ksmserver errors
If ksmserver fails to respond to a shutdown request it is more important
to continue the user's request of shutting down than it is to bail out.

This comes up more often the legitimate case wayland where an xwayland
crash would leave ksmserver not running, but does not log the user out.
We want to move in the direction of ksmserver being optional on wayland.

BUG: 449948