lookandfeel: Remove inconsistency between InputPanel::active and Qt.inputMethod.visible
Virtual Keyboard (InputPanel) would follow the input method's status
regardless of whether it visible on screen or not. In this case there
was an additional property which introduced a desync between those
states, and as a result virtual keyboard didn't show on a screen but
still consumed arrows and Enter key for its internal key navigation just
because input method "thinks" it is visible.

This patch removes the inconsistency, but the input panel would always
show up on start (on X11). So at least user can see what "eats" their
keys.