Fix systemd service file generation for cross-compiling
In case of cross-compiling, the qdbus version reported from qpath only
provides the path to the host's qdbus executable. For the binary that is
provided on the target, no path is available with the present
information.
For simplicity hardcode the qdbus path for this case. If every a
cross-compiled system emerges with non-standard /usr/bin/qtdbus making
this path settible via a build parameter might make sense.

Signed-off-by: Andreas Cord-Landwehr <cordlandwehr@kde.org>