dataengines/weather: Implement Warnings for NOAA provider
The NOAA service issues alerts for the US region, via the CAPv1.2
JSON API: https://www.weather.gov/documentation/services-web-api

The alerts are filtered by county because it includes all the events.
Using the forecast zone would miss some of them, and the lat/lon point
corresponds to the weather station, not necessarily the user location.

BUG: 370226
FIXED-IN: 6.1