applets/clipboard: fix tool buttons not clickable using stylus
Those buttons are exclusive grabbers for mouse events, but not for other
pointer events like stylus events, so those events are delivered to
the passive grabbers, and both the TapHandler and the ToolButton will
receive the same pointer event, but TapHandler has a higher priority and
when the TapHandler receives the event, the toolbutton actions become
invalid because the item has changed.

This disables the TapHandler when the toolbuttons are hovered to fix the
problem.

See also: https://bugreports.qt.io/browse/QTBUG-108821