Fix the translation templates (no version suffix for consistency)
While in other cases we have kept the version suffix in translation
templates, this has not been done for plasma modules.
As translation templates which use the "5" suffix should be fixed
anyway, follow the same pattern of the rest of Plasma modules and
just remove it instead of changing it to "6".