kcms/wallpaper: Add option to set the same wallpaper for all screens at once
Now that there is a centralized KCM for setting the wallpaper this seemed practical to do, so I gave it a go here. It just enumerates all screens using the same code as the `screenIdFromName` method and sets the wallpaper for each of them. I don't have additional displays, so I'm just using the virtual screen created by the screen sharing portal for testing.

FEATURE: 436001
FIXED-IN: 6.0