Increase default font size
From 10px to 11px. This is one of the biggest change, I do when
configuring a fresh Plasma session. It improves both the accessibility and
usability of Plasma by making the interface more readable. This is also
more in line with the font size used for websites which often even go up
to 14px (but we ihmo should no go that far).

People wanting a more compact interface can still change the font size
in the settings, which should not be an issue since these users tends to
more power users than users which want bigger font.