panelview: Set the correct mask
QWindow::mask sets the input regio, since the window contents
are shifted away from (0, 0) for top and left panels,
setting the mask without offset made the input region cover
only part of the panel.
Dont intersect the mask with the screenPane, it needs to be in window
local coordinates, geometryByDistance returns a rect in global coordinates.

BUG:479926
FIXED-IN:6.0