Restore Fitts' law for the panel
This is basically a revert of 21a1b654fe470437a4ff7954b370cf5a88a7c98d.
The justificaiton is correct the mask needs to be in local coordinates
this is/was achieved via mapFromGlobal. The problem was that at various
points the obtained geometry was intersected with the screen geometry.
Now that is is fixed we actually want to use screenPanelRect as the
input region which includes the area around the floating panel.
BUG:480016
FIXED-IN:6.0


(cherry picked from commit da2e891e740e65e0caffaa82735444419d3981ac)