Simplify logout greeter UI when invoked with a specific option
Right now the logout greeter always shows all available power/session
actions (except for Lock, for some reason) regardless of how it was
invoked, plus a cancel button and an OK button.

As a result, though it is essentially a confirmation screen, it offers the
user options to change their mind after the fact and perform a different
action from the one they started with, including those like Sleep and
Hibernate that don't exit the session at all.

This isn't ideal as it presents a somewhat strange and cluttered UI to
the user, full of options that are unlikely to be used.

To improve the situation, when the dialog is invoked with a specific
option, now we only show that option plus a cancel button.

When the dialog is *not* invoked with any option, (e.g. from the
off-by-default "Leave" item in the desktop context menu, or the
Ctrl+Alt+Delete global shortcut), all options are still shown, with
"Shut Down" pre-selected--as was the prior intention, but it wasn't
working. This required some re-plumbing as the dialog wasn't previously
designed to support this kind of dual-mode operation. The greeter's
windowed test mode is also updated to always show all available options.

BUG: 478022
FIXED-IN: 6.1