logout-greeter: Always answer dbus calls
Since "shutdown: Introduce support for KWin closing wayland windows"
plasma-shutdown is running for the whole time the KWin notification
is visible waiting for the user to choose whether to abort or
continue the logout.

If the user ignores the notification and clicks again to logout
we would hang the desktop while waiting for logout-greeter to answer.

BUG: 480264