kcms/wallpaper: Remove resolution line, use name of display
The KCM currently shows the name of the connector for the display. For
example, a monitor via DisplayPort is named DP-1.
Below that is the resolution of the display.

However, both of these are not really that useful to most people. They
are more likely to recognize the name of the brand than some
techno-term like "HDMI" or "DP". Not only that, but the resolution may
even be wrong because Qt gives us the wrong pixel size at fractional
scales. A 4K display at 125% reads back as 3072x1728, which is wrong.
This information isn't terribly useful here either.

BUG: 480291