kcms/soundtheme: Improve error message for sound not found
This error is the most common a user can experience in the KCM,
as it will happen if a theme doesn't provide one of the preview
sounds (ex. 'freedesktop' and sound 'batery-caution')

To minimize possible confusion for users (and bug reports), offer
a more clear and translated string on why it's happening, and
show it as an information message rather than an error.

CCBUG: 480626