shell: Reset activity switching shortcuts in plasma6.0-remove-old-shortcuts
Currently it's assumed that a kconf_update script is going to do this,
but the activity switching shortcut is not owned by kwin, so it should
not touch it. The kconf_update script in kwin is not correct too.


(cherry picked from commit 6601ce399e5a46f3cc0e4cbf41faee8179d45df9)