Don't remove add widgets/add panel from possible actions
Partly reverts 99f9fea6

That commit wanted to disable by default the actions "add panel" and "add widgets" from the desktop context menu.
Unfortunately, it removed those actions completely making it impossible to add them back from the config dialog.

To be disabled by default actions need to be inserted in the disabled array, but not be removed from m_actionOrder
as that removes it from the list of the existing available actions

BUG:480392