Revert "lookandfeel: Explicitely set isMask of ActionButton"
This reverts commit 2d46de674ced4062c53799772339ad152ffcbc69.

This change made sense to fix bug 474377 back when these icons came from
the Plasma style. However subsequently that feature was removed. As a
result, icons from the desktop icon theme are now shown here, and we
can't guarantee their visual style. As such, forcing them to be a mask
is now incorrect and obliterates the icon if it's using the full-color
style.

BUG: 477995
CCBUG: 474377
FIXED-IN: 6.0