panelview: Take focus from other Plasma windows when clicked
PanelView is a special case of a window, it is clickable, but it doesn't
take focus.

If we're interacting with other plasma components and a user clicks the
panel we still want focus to be drawn from those other components. This
patch makes PanelView fake focus out events. PanelView still does not
take focus, but any windows that close on focus out will work
appropriately.

BUG: 367815
FIXED-IN: 6.0