shell/panelview: calculate dodge geometry independent of floatingness
The dodge/defloat geometry should be the same no matter the panel is
floating or not. In addition, constrain it to the geometry of the
current screen, so that content from other screens have no impact.

Previously, the compensation for floatingness was done in Panel.qml,
which made the code more complicated, and had rounding errors that
caused the panel to continuously cycle between float and defloat.

CCBUG: 481379
FIXED-IN: 6.0