kcms/notifications: Fix delegate's minimum size
The RowLayout within the delegate will not shrink below its
content's implicit width, even when it overflows the external
width constraint.

We can use some Layout properties in a creative way to reduce
this, so that the TextField can shrink but keeps its regular
size when the space available is wider.

BUG: 481433
FIXED-IN: 6.0