libtaskmanager: find desktop file in cwd
For some Electron apps, the pid of the main window may not have enough
information to provide a desktop file due to the multiprocess architecture.
So try to find a desktop file with the same name in the same folder (or cwd),
and if there is a match, create a temporary service from the file to
read its icon.

BUG: 481490
FIXED-IN: 6.0