Reopen running apps on startup
On wayland we have no session restore that works. Gnome's implementation is
non-standard, there's nothing in Qt6. There is definitely nothing that
can be done in time for 6.0.

Even on X11 not all apps have good support, and with Flatpak requiring additional
permissions for XSMP it's getting worse.

So we need something to use as a fallback for both now, and for apps without support.

This is a simplistic solution that simply relaunch any wayland applications running at the time of shutdown.
X11 apps can restore as before through ksmserver as before.

It's not a smart solution, there's nothing that brings back the exact
same session or handles multiple windows.