applets/kicker: Always use DBus activation for launching apps
Setting the KIO::ApplicationLauncherJob::DeleteTemporaryFiles flag has
the effect of not using DBus activation for launching apps. This causes
subtly different launch and activation behaviors for when launching
apps from Kicker/Kickoff etc. compared to doing so using Task Manager,
KRunner, or all of the other app launching entry-points which do use
DBus activation. The result is unusual bugs, and inconsistent,
random-feeling behavior when things go wrong.

Let's always use the commmon DBus activation codepath here. If it's
broken or has bugs, we need to know about them and fix them, not provide
a hidden and unexpected workaround.

BUG: 479010
FIXED-IN: 6.1