Verify that install prefix is named "usr"
Because we use the parent dir of prefix as the image root
for systemd-sysext. So if it's named "kde", then systemd-sysext
won't find a "usr" dir in image, and consider the image invalid.

We can't build a image by symlinking the prefix to "usr" in
image dir, because systemd-sysext doesn't follow links inside image
for security concerns:
https://github.com/systemd/systemd/issues/31523

Also, moved the check for systemd to the beginning of the script,
so all checks are done before actually modify anything.