applets/systemtray: Prefer Icon over IconName
The systemtray has a custom KIconLoader in order to make the SNI's IconThemePath work.

When passing an icon name to Kirigami.Icon we skip that. Instead prefer the Icon role
from the model, which goes through the custom loader

This restores the logic to how it was before d31b985b6ba3d68b2ac018839864b305b6780574.

BUG: 479712
FIXED-IN: 6.0.1