kcm_fonts: fix combobox item not taking up full width
Added a width attribute to the combo boxes that evaluates to the
maximum of the container's width and its implicit width, so that
the items will take up the full space of its container even when
the preview images failed to render e.g. when it shortcuts under
Wayland.

Didn't expect it to interact well with the onWidthChanged handler,
but it magically worked.


(cherry picked from commit a54e6f9ee42ab1450cf2b9b89f1f8b1f12d1a11d)