session-restore: Avoid apps managed by ksmserver
If a session is saved manually plasma-fallback-session-restore and
ksmserver can contain the same entries. To avoid double-starting, the
fallback-session-restore introspects the ksmserver config and skips
accordingly.

This also migrates to kcfg to keep future defaults in sync easier.

It's not the best architecture, but it's intended as a fallback mechanism.

CCBUG: 4009