libkmpris: prevent an empty icon from being used
The previous logic would only use the fallback icon
("emblem-music-symbolic"), if the client had the "DesktopEntry"
property[1] set.

But since clients are allowed to omit that property completely
(mpd-mpris[2] does that), m_iconName would stay an empty string, and
clients would be left icon-less.

This change ensures that m_iconName is never an empty string.

BUG: 482908

[1] https://specifications.freedesktop.org/mpris-spec/latest/Media_Player.html#Property:DesktopEntry
[2] https://github.com/natsukagami/mpd-mpris/