kcms/nightcolor: Make manual lat/lon respect locale decimal point
Use the locale-aware parsing functions for the decimal values

To be able to work with the TextField validator in a semi-sane
way without handicapping the user input, just update the backend
value when finishing the edition.

This let's us simplify the logic and also fix resetting to the
last good value on a wrong input, which was intended but also
wasn't working before.

BUG: 413956
FIXED-IN: 6.0.3