shell: emit screenGeometryChanged when panel is ready
When screenGeometryChanged is emitted in `handleScreenOrderChanged`,
only desktops are ready. Panels are still in the waiting list and do not
have a associated containment, so those containments do not have a valid
id for screen, which is required to get a valid screen geometry.

BUG: 482339
BUG: 481736
FIXED-IN: 6.0.3


(cherry picked from commit 6d543aca43858f707b32e1e08c39bfd7b3ffaf8e)