applets/appmenu: fix keeping button always visible in compact mode
In compact mode, the button tries to disable itself rather than hiding
when the active app doesn't support the global menu, but this gets
undone by the widget setting itself to hidden when in that state.

We can minimally fix this by adding an extra condition to the status
property code.

BUG: 435928
FIXED-IN: 6.0.3