devicenotifications: Actually remember device name
On disconnect, device properties are already gone, so we need
to remember the device name for the notification. The infra for
doing that was all there but the name was never actually remembered.

Must have been a bad rebase.


(cherry picked from commit 368d7036f6c558c8d37e147b29cba59815b8c629)