applets/digital-clock: Simplify binding for hideOnWindowDeactivate
This is the `root` context already, no need for a Binding object. This
is probably an artifact from the refactoring where some of the similar
bindings were located in other contexts and were bound to the
AppletInterface object (via Plasmoid.self attached property) rather
than via the new PlasmoidItem API. Just a guess though.