applets/digital-clock: Rename nameForZone method for clarity
The nameForZone name did not make any sense. First, the terminology is
bad: it's a "time zone", not just any "zone". Second, it does not
return a name of a time zone -- it, in fact, takes such name as a
parameter, and returns an appropriate user-facing string representing a
time zone based on current plasmoid configuration.