applets/digital-clock: Improve last selected time zone matching
Previously, if user has manually added their local time zone to the list
of selected time zones, and selected specifically it as their primary
one, they would not get bold text highlight in the calendar pane nor in
a tool tip, because the condition only strictly matched on the
modelData and lastSelectedTimezone. Instead we want to compare resolved
names, because Plasmoid.configuration.lastSelectedTimezone will contain
a particular time zone code (such as "Asia/Almaty" which might resolve
to simply "Almaty") while the model for the UI will contain the "Local"
time zone as per logic in the
selectedTimeZonesDeduplicatingExplicitLocalTimeZone method which might
happen to resolve to the same "Almaty" city name.