components/calendar: DaysModel: Optimize QHash and QMultiHash queries
QHash::count is basically 0 or 1 depending on the result of a boolean
contains() query, so it doesn't make sense twice as much, because the
int would later be cast back to bool for the `if` condition.

QMultiHash::count is more optimal than values().count() because it does
not need to build a throw-away list.