components/calendar: Reduce visual glitches when resizing the view
This is still not ideal, but I guess it depends on the non-deterministic
order or property change signal propagation -- the actual content view
might adjust cell dimensions and resize before or after the SwipeView.
Note that binding on contentWidth & contentHeight changes doesn't help.
Ideally, they would've communicated via a protocol with well-defined
order and semantics.