kcms/nightcolor: require explicitly consenting to auto location
Otherwise the user will immediately be located as soon as they choose the current location mode.
Since we want to move the enabled/disabled control from a pseudo "switching times" mode into a dedicated switch and autolocation is the default switching times mode, enabling night light would immediately start the locator, so we need to make sure the user consents to this.