kcms/nightlight: Drop icon theme hack, switch map zooming icons to the new lightweight +/- symbols
Even with the new magnifier icons the buttons looked infinitely worse
than any other alternative. That's not to say the icons are bad, they
are just not suitable for zooming maps imho. So we introduced simple
zoom-plus & zoom-minus icons like every other GIS software out there.

They are links to list-add and list-remove because we want them to look
like a simple + and - signs, which appears to be a widespread
convention in most GIS software. We don't want maps to have zoom icons
with a magnifying glass because a magnifying glass is already used for
search buttons/fields which are also fairly common in such apps.

While at it, switch to symbolic versions.