kcms/users: remove clear all fingerprints button
It is very easy to accidentally click on the clear all fingerprints button,
resulting in the loss of all fingerprints, which can take long minutes to
re-enroll. We can have only 10 fingerprints, and each one has a delete
button, so it is not necessary to have a "clear all" button.