applets/brightness: fix night light inhibition state not synced across instances of the applet
Integrate inhibition properties into the monitor class so it becomes a night light control class, which also simplifies the back-of-the-frontend code.

Then use a static self member for the night light inhibitor from the control class, so that the inhibition state data (state, pending inhibits and cookie) can be shared among the applet instances.

Monitor the inhibited property to also pick up inhibitions created from outside the applet.

BUG: 438291