applets/brightness: only attempt to release our own night light inhibitions
We can only release night light inhibitions that we ourselves (i.e. the applet) created as we need to pass a matching cookie; this is a restriction of the DBus API. If we attempt to release an inhibition not created by ourselves (e.g. by the keyboard shortcut) by passing a wrong cookie it will fail, so make the inhibition switch checked but disabled in that case.