[widgetexplorer] Simplify icon lookup
Move all of the lookup into the ctor.

If the metadata contains an icon name that starts with a / look up the
icon in the KPackage, if not take the icon from the theme

If no icon name is given check whether the applet name is an icon,
otherwise use a generic fallback

Always give the icon name as data, that way the view can load the icon
with the correct colors. This fixes icon colors with dark Plasma/light
application theme

Co-authored-by: Evgeniy Harchenko <evgeniyharchenko.dev@gmail.com>