kcms/icons: filter out GNOME's Adwaita and High Contrast icon themes
These are no longer FDO-compatible icon themes, and are apparently no
longer intended to be used for non-GNOME apps--going against what an
FDO icon theme is supposed to be used for.

I've contacted the GNOME folks about it, and unfortunately they've made
it clear that the situation is intended; see
https://gitlab.gnome.org/GNOME/adwaita-icon-theme/-/issues/288

As such, allowing the user to select these icon themes will just give
them an opportunity to break all their non-GNOME apps. Let's filter
these themes out to prevent that possibility.

We still need to figure out a solution for when our apps are run in
distros where Adwaita is used as the default icon theme, but that's a
separate topic.

BUG: 486409
FIXED-IN: 6.1