kcms/lookandfeel: Make it super duper obvious what layout replacement does
This action will destroy all your shell customizations and replace them
with what's in the theme. The way we express this is a bit less clear
than it needs to be IMO; I worry that we're not properly communicating
the dangerousness of this option.

As such, re-word the warning text to make the dangerousness clear.