Adopt a merge request template
The idea here is to encourage people who write commit messages that
don't adequately explain relevant information to at least clean up
the merge request description, which makes their merge requests more
likely to be reviewed and makes life easier for reviewers. At
merge-time, the MR can be suqash-merged to make the final commit
message echo the MR description.

For people who already write good commit messages or send MRs of
multiple curated commits, nothing really changes.

This has worked out very well for Elisa, which has used this template
for a year. It's modeled on the template we used for Phabricator.
Copy it here and give it a try for Plasma.

If it works out, I'd eventually like to do a sysadmin thing that
automatically syncs this into every repo that opts in, so we don't
have to copy it around. But at this point I think that's a bit
premature until we get more widespread usage and acceptance.