Notifications: Restore custom DragHelper
When the notification popup closes, the ListView mode is unloaded to
save some resources. On Wayland, during drag and drop another window
might be raised and focussed causing the notification popup to close.

QtQuick's built-in drag handling does not take kind to having the
item it's operating on destroyed behind its back, particularly because
it uses a freaking' nested event loop inside of dragStarted!

Restore the original DragHandler which detaches the drag operation
entirely from the item that triggered it. It also creates a proper
drag pixmap for dragging the file icon and keeps it for the entire
duration of the drag.

As a bonus this also lets us export the URL to the Document portal
to enable drag and drop of thumbnails to sandboxed apps.

BUG: 484546
BUG: 486665