xembedsniproxy: Set input region when hidden to avoid grabbing clicks
xembedsniproxy is a hidden window. It is used for grabbing content and
replaying events.
It tried to hide itself by adjusting the stacking order.

This is insufficient on wayland where we might be the lowest X window,
but that can still be above the desktop. Instead set an empty shape
when we should be hidden.

BUG: 433079