Zoom out when going in edit mode
Makes sure the whole desktop area is reachable by zooming it out and putting
controls around it. Look and feel is similar with the KWin Overview effect.

The panel edit window tries to be right-aligned to leave as much space as
possible for the zoomed out desktop window.

Other part is https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2235