xembedsniproxy: Set input region when hidden to avoid grabbing clicks
xembedsniproxy is a hidden window. It is used for grabbing content and
replaying events.
It tried to hide itself by adjusting the stacking order.

This is insufficient on wayland where we might be the lowest X window,
but that can still be above the desktop. Instead set an empty shape
when we should be hidden.


### Test plan

Had a floating compact panel (this is important as it means the system tray icon moves about when I open new windows)

Ran hexchat and clicked on the icon and and other icons.

xembedsniproxy was compiled with "VISUAL_DEBUG" flag enabled so I could see where to click, and now (with XWayland >= 2.41) events are never intercepted from the panel.

<!--
If this merge request introduces a visual change, please add before-and-after screenshots in the following format:

| Before | After |
| ------ | ----- |
| [drag "before" screenshot here] | [drag "after" screenshot here] |
-->

### Bugs fixed

BUG: 433079
<!--
If the changes in this merge request fix any Bugzilla tickets, add the following keyword for each one:

BUG: [number of the bug report]
-->