weather/noaa: Improve the days shown in the forecast
Previously, a day would be filtered out from showing in the
applet when it didn't have for instance the minimum temperature
(which usually happens for the last day)

Let's filter out day forecasts only if they don't provide the
weather condition, instead of min/max temperatures.

Also report the actual number of days after the filtering,
instead of the total days in the report. Otherwise we could
show less days than we claim on the tab.