shell: Leave edit mode if show desktop is externally disabled
When we enter edit mode we enable show desktop to move all windows out
the way and change the stacking order of the desktop view. If this is
externally manipulated we want to leave edit mode so the user is reset
to a sane state.


(cherry picked from commit aa7fce142b36ca4588a971aebb01b03b0b42a81b)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>