Remove iceauth dependency
Previously the tool iceauth was used to write authentication to and from
the iceauthority file, running a set of commands to manipulate the file.

This external dependency is not often installed, which may create a
problem
now that all paths go through authentication. If iceauth is not
currently installed
ksmserver will assert rendering the session useless.

This patch writes and manages the file directly removing the need for
external tools.

Co-authored-by: Fabian Vogt <fabian@ritter-vogt.de>


(cherry picked from commit 9398f6cf8933055b31506ba155aef2fc2b3561d7)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>