colors/colorsapplicator.cpp: Try to compensate for low contrast ratio for foregroundLinks
If foreground link contrast ratio with the background is too low, try to
make it a bit brighter. It will not always work, but it's better than
nothing. 

This can be especially noticeable with the "Accent color from wallpaper" setting.

It's not perfect, but it does help a lot with readability.