weather/noaa: Fix the URL for retreving the station list
It seems the noaa outages on the legacy API continue, and the
previous URL that provided the list of weather stations is no
longer working (it errors out as a 404).

There is however a probable alias/mirror that still works
(`w1.weather.gov`).

Instead of a pure replacement, add a small fallback mechanism
to also check the older URL in case it changes again in the
future. At some point we might be even want to cache it in a
local file.


(cherry picked from commit 386985ea5f5c5d2ff6ce981092d70f2a38f1f7f9)

Co-authored-by: Ismael Asensio <isma.af@gmail.com>