Don't cause unnecessary resizes on containment change
When an applet is moved from a contianment to another, for a moment it
was parented to the whole AppletsLayout.

since it had anchors.fill:parent set, it was resized to the whole
appletslayout geometry or its max size if had any. for instance the
kickoff applet would be resized to the whole screen

### Reason for the change

now applet size shouldn't change anymore when draggin from on e screen
to the other

### Test plan

drag the kickof applet from a deskop to the one of a different screen