Don't cause unnecessary resizes on containment change
When an applet is moved from a contianment to another, for a moment it
was parented to the whole AppletsLayout.

since it had anchors.fill:parent set, it was resized to the whole
appletslayout geometry or its max size if had any. for instance the
kickoff applet would be resized to the whole screen

(cherry picked from commit b3a5eff557926ee113f292836ef3a3c5d8131a30)

511616b5 Don't cause unnecessary resizes on containment change

Co-authored-by: Marco Martin <notmart@gmail.com>