startkde: forward sigterm to child process in dev setup scripts
When a bash script is SIGTERM'ed it will not forward the signal to any
children. Instead they get a SIGHUP.

It's important that startplasma gets the right signals from the login
manager for teardown cases.

We cannot use 'exec startplasma' as there is a cleanup script.

This change only affects developers and requires the user to call
./install-sessions.sh to apply.