startkde: Close systemd session when startplasma exits
Startplasma is the session leader managed by the display manager,
Anything started by systemd needs tearing down explicitly.

On logout or shutdown we correctly stop the graphical-session target and
then startplasma quits.

If SDDM is restarted the order happens in reverse. startpasma quits, but
we also need our units to close. Otherwise we end up in a situation
where we're at the login manager but with all our background services
continuing to run.

Given startplasma starts the plasma-workspace target it's our
responsibility to make sure this target gets stopped.

This is an alternative to 0ff2e9a70ccfa96b82fd6d2ea5a0a8545a5e3a60 but
considerably simpler. It doesn't handle the case of startplasma crashing
or being SIGKILLED, but is considerably simpler.